requirements.txt
# ============================================
# SGG-API - Dependencias de Producción
# ============================================

# FastAPI Framework
fastapi==0.128.0
uvicorn[standard]==0.40.0
starlette==0.50.0

# Pydantic (Validación de datos)
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
annotated-types==0.7.0
typing-extensions==4.15.0
typing-inspection==0.4.2

# Database
SQLAlchemy==2.0.46
pymysql==1.1.2
greenlet==3.3.1
alembic==1.18.1
Mako==1.3.10

# Authentication & Security
python-jose==3.5.0
passlib==1.7.4
bcrypt==5.0.0
cryptography==46.0.3
cffi==2.0.0
ecdsa==0.19.1
rsa==4.9.1
pyasn1==0.6.2

# File Upload
python-multipart==0.0.22

# Validation
email-validator==2.3.0
dnspython==2.8.0
idna==3.11

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
annotated-doc==0.0.4

# File handling
Pillow==12.1.0
reportlab==4.4.9
charset-normalizer==3.4.4

# Email
aiosmtplib==5.1.0

# Environment variables
python-dotenv==1.2.1

# HTTP Client
httpx==0.28.1
httpcore==1.0.9
certifi==2026.1.4
anyio==4.12.1

# Excel/CSV
openpyxl==3.1.5
pandas==3.0.0
numpy==2.4.1
et-xmlfile==2.0.0
tzdata==2025.3

# JSON
orjson==3.11.5

# Server (Uvicorn dependencies)
click==8.3.1
h11==0.16.0
httptools==0.7.1
pyyaml==6.0.3
watchfiles==1.1.1
websockets==16.0
colorama==0.4.6

# Security
MarkupSafe==3.0.3
pycparser==3.0